--- a/src/main.py
+++ b/src/main.py
@@ -2,6 +2,9 @@
 from os import path
 
 import sys
+# Workaround for kodi RuntimeError https://kodi.wiki/view/Python_Problems#asyncio
+sys.modules['_asyncio'] = None
+import asyncio
 from elementum.provider import register
 from kodi_six import xbmcgui
 
@@ -9,6 +12,11 @@ from logger import log
 
 sys.path.insert(0, path.realpath(path.join(path.dirname(__file__), '..', 'resources', 'libs')))
 sys.path.insert(0, path.dirname(__file__))
+
+# fix asyncio.run() on win
+if sys.platform == "win32" and (3, 8, 0) <= sys.version_info < (3, 9, 0):
+    asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
+
 
 if __name__ == '__main__':
     import debugger
@@ -26,7 +34,7 @@ if __name__ == '__main__':
         sys.exit(1)
 
     if sys.argv[1] == "validate_settings":
-        jackett.validate_client()
+        asyncio.run(jackett.validate_client())
     else:
         debugger.load()
         register(