--- a/src/utils.py
+++ b/src/utils.py
@@ -116,3 +116,11 @@ def get_provider_color(provider_name):
         colors[i] = f'{colors[i]:02x}'
 
     return "FF" + "".join(colors).upper()
+
+
+def check_season_name(title, season_name=""):
+    # make sure season name is unique. Not eq to movie title or "season". It saves from false-positive filtering.
+    season_name = season_name.lower()
+    if season_name in title or "season" in season_name:
+        return ""
+    return season_name