Big QOL PR. 
- Speed up
- More info on the screen
- Separated caps for indexers
- Only one thread compere to old one. Coroutines used here. 
- Smart filter for TV search. Works much accurate but slower. Works better with https://github.com/elgatito/elementum/pull/86 and https://github.com/elgatito/elementum/pull/87

**Smart Filtering compare to episode + season search:**
Smart Search "game of thrones S04E06" :
```
##  request show by name only. So filtering all possible torrents.
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutor/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutracker/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/kinozal/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/lostfilm/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/megapeer/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anilibria/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anidub/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones'}
[script.elementum.jackett] 272 ... filtering resolution                        ## 272 torrents returned by all indexers
[script.elementum.jackett] 272 ... filtering release type
[script.elementum.jackett] 272 ... filtering no seeds
[script.elementum.jackett] 272 ... smart-filtering show torrents
[script.elementum.jackett] 60 ... resolving unique magnets                     ## got 60 torrents after filtering
[script.elementum.jackett] 20 torrents returned from Jackett in 21.05 seconds  ## 20 is default limit
```
![image](https://github.com/fugkco/script.elementum.jackett/assets/43271561/d1d43e75-5be3-4b10-95bf-7d5b2b784309)
episode + season Search "game of thrones S04E06":
```
## requestin by season+ep and by season
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/lostfilm/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/kinozal/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anilibria/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutor/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutracker/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/megapeer/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/lostfilm/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/kinozal/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutracker/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutor/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anilibria/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anidub/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/megapeer/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4}
[script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anidub/results/torznab params: {'t': 'tvsearch', 'q': 'game of thrones', 'season': 4, 'ep': 6}

[script.elementum.jackett] 91 ... filtering resolution
[script.elementum.jackett] 91 ... filtering release type
[script.elementum.jackett] 91 ... filtering no seeds
[script.elementum.jackett] 59 ... resolving unique magnets
[script.elementum.jackett] 20 torrents returned from Jackett in 15.06 seconds
```
![image](https://github.com/fugkco/script.elementum.jackett/assets/43271561/06af6ecd-e08f-4c38-b748-1b8cc6654cd7)
you can see the first one is found only because of 4K have 4 in in name  like a season we are looking for.
Smart filtering is more relevant. 
**List of indexers now requested before search. Search each indexer separately in async mode.**
```
## geting indexers list
2023-11-03 00:53:49.041 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/all/results/torznab?t=indexers&configured=true params: None
2023-11-03 00:53:49.203 T:18152    info <general>: [script.elementum.jackett] Jackett returned response in 0.156s size 110722 b

## requesting indexers "q" used because of "smart-search"
2023-11-03 00:53:49.228 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anilibria/results/torznab params: {'q': 'game of thrones'}
2023-11-03 00:53:49.229 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutracker/results/torznab params: {'q': 'game of thrones'}
2023-11-03 00:53:49.230 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/rutor/results/torznab params: {'q': 'game of thrones'}
2023-11-03 00:53:49.230 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/anidub/results/torznab params: {'q': 'game of thrones'}
2023-11-03 00:53:49.231 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/kinozal/results/torznab params: {'q': 'game of thrones'}
2023-11-03 00:53:49.232 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/megapeer/results/torznab params: {'q': 'game of thrones'}
2023-11-03 00:53:49.233 T:18152    info <general>: [script.elementum.jackett] Sending to Jackett url: /api/v2.0/indexers/lostfilm/results/torznab params: {'q': 'game of thrones'}

##got responses
2023-11-03 00:53:51.743 T:18152    info <general>: [script.elementum.jackett] Jackett returned response in 2.500s size 95463 b
2023-11-03 00:53:51.774 T:18152    info <general>: [script.elementum.jackett] Done searching. Got 50 torrents from Kinozal.
2023-11-03 00:53:52.262 T:18152    info <general>: [script.elementum.jackett] Jackett returned response in 3.016s size 126287 b
2023-11-03 00:53:52.308 T:18152    info <general>: [script.elementum.jackett] Done searching. Got 50 torrents from MegaPeer.
2023-11-03 00:53:52.762 T:18152    info <general>: [script.elementum.jackett] Jackett returned response in 3.531s size 53288 b
2023-11-03 00:53:52.778 T:18152    info <general>: [script.elementum.jackett] Done searching. Got 22 torrents from LostFilm.tv.
2023-11-03 00:53:55.986 T:18152    info <general>: [script.elementum.jackett] Jackett returned response in 6.750s size 511 b
2023-11-03 00:53:55.999 T:18152    info <general>: [script.elementum.jackett] Done searching. Got 0 torrents from AniDUB.
2023-11-03 00:53:59.930 T:18152    info <general>: [script.elementum.jackett] Jackett returned response in 10.688s size 167582 b
2023-11-03 00:53:59.963 T:18152    info <general>: [script.elementum.jackett] Done searching. Got 50 torrents from RuTracker.
```